Crisis RLOC for all assessment types
Lately staff have been scoring the risk scores higher on updates,
because the crisis rloc was only being returned on intake/crisis the
high score was allowed to be saved that would later cause an error in
CMBHS. Crisis will be returned for any assessment type now. The
assessment cannot be saved of the type is 1,2 or 3 because of another
validation about the client already being in a package.